requirements.txt
appdirs==1.4.4
asgiref==3.3.1
autopep8==1.5.2
distlib==0.3.0
dj-static==0.0.6
Django==3.1.5
django-ckeditor==5.9.0
django-cleanup==4.0.0
django-countries==6.1.2
django-crispy-forms==1.9.0
django-etc==1.3.0
django-hitcount==1.3.3
django-js-asset==1.2.2
django-recaptcha==2.0.6
filelock==3.0.12
iyzipay==1.0.32
Pillow==7.1.2
pycodestyle==2.5.0
pytz==2019.3
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
virtualenv==20.0.21
wfastcgi==3.0.0
